src/requirements.txt
flask==3.1.2
requests=2.32.5
pyyaml==6.0.3
dotenv==0.9.9